requirements.txt
pandas==1.0.4
numpy==1.18.4
scipy==1.4.1
matplotlib==3.2.1
scikit-learn==0.23.1
scikit-optimize==0.8.1
sympy==1.6.2
pyspark==3.0.1
jmetalpy==1.5.5
Flask==1.1.2
Flask-Compress==1.5.0
Flask-Cors==3.0.8
future==0.18.2
docutils==0.16
plotly==4.10.0
chart_studio==1.1.0
dash_bootstrap_components==0.10.7
flask_caching==1.9.0